Added enter committee transaction with assertion
